Enable catchments for searches in Tasmania
Some searches for services in Tasmania return services which are
located in Victoria because those services happen to be physically
close to the the Tasmanian location. To prevent that, enable
catchments for any searches in Tasmania.

Specifically, searching for services near Burnie, Tasmania
returns some services in Melbourne and other parts of Victoria.

We may extend this to other states as catchment data is added.