Don't inform Google Analytics of the details of searches
If the user has performed a free text search, the URL will include
the search terms. We don't want to inform Google Analytics of the
content of the user's search, because chances are that GA can
link that to the user's Google profile, despite whatever we put in
place to "anonymize" the data collected by GA, and we don't want
Google knowing that a person is, e.g., fleeing from domestic
violence, in case that information somehow leaks out.