Update the environment variables when `serve` is run
`serve` may be run without `deploy` being run first, so we need to
ensure that the environmnet variables are always up to date when
`serve` is run.

For instance, when the docker container's environment variables
are updated, the container is restarted and the `serve` command
is run, without the `deploy` command being run at all. This means
that the container's new environment variables will not be
reflected in the application code.