Set up Google Analytics and Tag Manager in prod environments
The previous method of initialising Google Analytics and Tag
Manager didn't play well with the fact that we're now pre-rendering
everything and serving it as static files.

Instead, we just include a link to the static analytic setup code
which is completely separate from any of the React rendering
pipelin. This simplifies things considerably, despite being a bit
un-idiomatic.

Conflicts:
	invoke.sh