Query SHS shelters when accessing housing through the footer
When accessing housing going through the shelter footer -> Find a specialist homeless service (link) it will now only fetch the SHS shelters. This is done by passing a link state flag to the route and overriding the current housing query.
The link in the CMS will need to change and be replaced with
```
{pathname: "/housing", state: { "shs": true }}
```