Fix resource leak with proxy
The external resources proxy uses the http-proxy-middleware library to
mange the proxing of requests. That in turn uses the node-http-proxy
library which unfortunately has not been updated in a while and has
a bug which results in sockets not being closed properly. This commit
fixes that by swapping that library out with a patched version found
here https://github.com/http-party/node-http-proxy/pull/1559