Remove tests of NODE_ENV for checking if test
Using NODE_ENV to check if the environment is a test environment or not
can be misleading since Next.js will automaticity override NODE_ENV on the client-side to be either production or development (depending on
which server is used). The value "test" won't be overridden server-side
but it can be easy to get confused when the server-side and client-side
use different values so best avoid that by using a different env var to
check if we're in a test environment or not.